Added ResultTests
